fix(python sdk): moved gemini client to google.py (#865)
Moved gemini_client to google.py and deleted gemini.py

---------
